Use streamr.network transport
Use streamr.network transport to implement pubsub chat.

Main changes:
* Support only arm64-v8a and x86_64 ABIs
* Add JNI module for streamr.network
* Add streamr.network interface modules
* New libnode.so modules for arm64-v8a and x86_64 based on node.js 16.3.12
* Updated node.js include files
* New nodejs-mobile modules
* Add gradle task to build JavaScript zip-file
* streamr-client node_modules in lib/src/main/js
* JavaScript interface code to streamr-client

Note that the name and authentication information for streamr.network
STUN/TURN servers is not exposed in this commit. Information can be
requested from streamr.network. Also the address of the public stream
is hidden in this commit. These settings are in the file:

lib/src/main/js/lib/api-controller.js

Signed-off-by: Markku Kylänpää <markku.kylanpaa@vtt.fi>